+++ code/src/app/[lang]/registry/task/closed/page.tsx
@@ -5,6 +5,7 @@ import { Modules, PageTitles, Routes } from 'io/config/routes';
 import SettingsStore from 'io/config/settings';
 import { NavBarItemSettings, TableColumnOrderSettings, UISettings } from 'types/settings';
 import RegistryTableComponent from 'ui/graphic/table/registry/registry-table-component';
+import { LifecycleStageMap } from 'types/form';
 
 /**
  * Set page metadata.
@@ -31,7 +32,8 @@ export default function RegistryTaskByDatePage() {
     return (
       <RegistryTableComponent
         entityType={uiSettings.resources?.registry?.data}
-        lifecycleStage={'closed'}
+        lifecycleStage={LifecycleStageMap.CLOSED}
+        accountType={uiSettings.resources?.billing?.paths?.find(path => path.type === LifecycleStageMap.ACCOUNT).key}
         tableColumnOrder={tableColumnOrderSettings}
       />
     );